+++ src/main/php/web/auth/oauth/BySignedRequests.class.php
@@ -1,21 +1,26 @@
 <?php namespace web\auth\oauth;
 
-use peer\http\HttpConnection;
-
-class BySignedRequests implements Client {
+class BySignedRequests extends Client {
   private $signature;
 
+  /** Creates a new instance with a given OAuth1 signature */
   public function __construct(Signature $signature) {
     $this->signature= $signature;
   }
 
-  public function fetch($url, $params= []) {
-    $c= new HttpConnection($url);
-    return new Response($c->get($params, [
+  /**
+   * Authenticates request and returns it
+   *
+   * @param  peer.http.HttpRequest $request
+   * @return peer.http.HttpRequest
+   */
+  public function authenticate($request) {
+    $request->addHeaders([
       'Accept'        => 'application/json',
       'User-Agent'    => 'XP/OAuth1',
-      'Authorization' => $this->signature->header('GET', $url, $params),
-    ]));
+      'Authorization' => $this->signature->header($request->method, $request->url->getURL(), $request->parameters)
+    ]);
+    return $request;
   }
 
   /** @return web.auth.oauth.Signature */